+++ config/settings/production.py
@@ -1,5 +1,4 @@
 import ldap
-
 from django_auth_ldap.config import LDAPGroupQuery, LDAPSearch, PosixGroupType
 
 from .base import *  # noqa
@@ -10,7 +9,7 @@ from .base import env
 # https://docs.djangoproject.com/en/dev/ref/settings/#secret-key
 SECRET_KEY = env("DJANGO_SECRET_KEY")
 # https://docs.djangoproject.com/en/dev/ref/settings/#allowed-hosts
-ALLOWED_HOSTS = env.list("DJANGO_ALLOWED_HOSTS", default=["radical_translations.info"])
+ALLOWED_HOSTS = env.list("DJANGO_ALLOWED_HOSTS", default=["radicaltranslations.org"])
 
 # DATABASES
 # ------------------------------------------------------------------------------
@@ -82,7 +81,7 @@ TEMPLATES[-1]["OPTIONS"]["loaders"] = [  # type: ignore[index] # noqa F405
 # https://docs.djangoproject.com/en/dev/ref/settings/#default-from-email
 DEFAULT_FROM_EMAIL = env(
     "DJANGO_DEFAULT_FROM_EMAIL",
-    default="Radical Translations <noreply@radical_translations.info>",
+    default="Radical Translations <noreply@radicaltranslations.org>",
 )
 # https://docs.djangoproject.com/en/dev/ref/settings/#server-email
 SERVER_EMAIL = env("DJANGO_SERVER_EMAIL", default=DEFAULT_FROM_EMAIL)