+++ radical_translations/contrib/sites/migrations/0003_set_site_domain_and_name.py
@@ -13,7 +13,7 @@ def update_site_forward(apps, schema_editor):
     Site.objects.update_or_create(
         id=settings.SITE_ID,
         defaults={
-            "domain": "radical_translations.info",
+            "domain": "radicaltranslations.org",
             "name": "Radical Translations",
         },
     )