+++ CHANGELOG.rst
@@ -21,7 +21,7 @@ Added
 Changed
 ~~~~~~~
 * Do not display French Republican dates by default.
-* Domain name.
+* Domain name, radicaltranslations.org.
 
 Fixed
 ~~~~~