+++ CHANGELOG.rst
@@ -14,6 +14,9 @@ The format is based on `Keep a Changelog`_, and this project adheres to
 [Unreleased] - yyyy-mm-dd
 --------------------
 
+Fixed
+~~~~~
+* Update Django Controlled Vocabulary app.
 
 [0.6.2] - 2020-06-02
 --------------------