+++ CHANGELOG.rst
@@ -20,9 +20,21 @@ Added
 * `Agent` views.
 * `Event` views.
 
+[0.6.3] - 2020-06-08
+--------------------
+
+Changed
+~~~~~~~
+* `Date` display format to include radical date when available.
+* Prefix paratext `Resources` with `[paratext]`.
+* Replace `Classification` `source` with editorial classification field.
+
 Fixed
 ~~~~~
 * Update Django Controlled Vocabulary app.
+* Add missing vocabulary entry for Printing and Publishing Terms.
+* Autocomplete for `Event` and `Place` models.
+* Issue deleting `Resource` contributions.
 
 [0.6.2] - 2020-06-02
 --------------------